implement DefaultDeviceSettingsModal
