+++ .cspell.json
@@ -48,6 +48,7 @@
     "roadmap",
     "roboform",
     "schemaname",
+    "signup",
     "sleaper",
     "stackoverflow",
     "subfolders",
