+++ mobile-app/package.json
@@ -76,6 +76,7 @@
     "jsc-android": "^250231.0.0",
     "jssha": "^3.3.0",
     "jwt-decode": "^3.1.2",
+    "lodash": "^4.17.21",
     "make-plural": "^7.3.0",
     "ms": "^2.1.3",
     "native-base": "^3.4.28",
